locust/requirements.txt
locust
websocket-client